#fix 462 add implement for RewriteFiles support Hive
